renamed complexity analysis file structure to distinguish this feaure as being type complexity only
